LabsPBZ/5sem/SecondLab/SecondLab/requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
pandas>=2.0.0

